fix warnings
git-svn-id: http://svn.osgeo.org/qgis/trunk@9594 c8812cc2-4d05-0410-92ff-de0c093fc19c

+++ src/core/raster/qgsrasterlayer.cpp
@@ -1228,7 +1228,7 @@ void QgsRasterLayer::computeMinimumMaximumEstimates( QString theBand, double* th
  */
 QgsContrastEnhancement* QgsRasterLayer::contrastEnhancement( unsigned int theBand )
 {
-  if ( 0 < theBand && theBand <= ( int ) bandCount() )
+  if ( 0 < theBand && theBand <= bandCount() )
   {
      return &mContrastEnhancementList[theBand - 1];
   }
